requirements.txt
ultralytics
gradio
opencv-python
numpy
Pillow
moviepy